requirements.txt
python-telegram-bot
requests
bs4